# docs/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Token API
  version: 1.0.0
  description: Версионированный API платформы для репетиторов и учеников.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /me:
    get:
      tags: [Identity]
      operationId: getCurrentUser
      responses:
        '200':
          description: Текущий пользователь и профиль
          headers: { ETag: { schema: { type: string } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/CurrentUser' } } }
        '401': { $ref: '#/components/responses/Problem' }
...
  /screens/{screen}:
    get:
      tags: [Screens]
      operationId: getScreen
      parameters:
        - in: path
          name: screen
          required: true
          schema: { $ref: '#/components/schemas/ScreenName' }
      responses:
        '200': { $ref: '#/components/responses/ScreenState' }
        '404': { $ref: '#/components/responses/Problem' }
...
  /lessons/{lessonId}:
    get:
      tags: [Scheduling]
      operationId: getLesson
      parameters:
        - { in: path, name: lessonId, required: true, schema: { type: string } }
      responses:
        '200':
          description: Занятие
          headers: { ETag: { schema: { type: string, example: '"v3"' } } }
          content: { application/json: { schema: { type: object, additionalProperties: true } } }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Scheduling]
      operationId: updateLesson
      parameters:
        - { in: path, name: lessonId, required: true, schema: { type: string } }
        - { in: header, name: If-Match, required: true, schema: { type: string, example: '"v3"' } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { type: string, enum: [planned, done, moved, cancelled, missed] } }
      responses:
        '200': { $ref: '#/components/responses/CommandSuccess' }
        '412': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
  /read-models/{model}:
    get:
      tags: [Screens]
      operationId: getReadModel
      parameters:
        - in: path
          name: model
          required: true
          schema: { type: string, enum: [student-dashboard, tutor-today, assignment-progress, student-subject-stats] }
        - { in: query, name: student, schema: { type: string } }
      responses:
        '200':
          description: Проекция, построенная из первичных фактов
          headers: { ETag: { schema: { type: string } } }
          content:
            application/json:
              schema:
                type: object
                required: [items, updatedAt]
                properties:
                  items: { type: array, items: { type: object, additionalProperties: true } }
                  updatedAt: { type: string, format: date-time }
        '404': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: token_sid
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, instance, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors: { type: array, items: { type: object } }
...
    CurrentUser:
      type: object
      required: [user, home]
      properties:
        user: { type: object, additionalProperties: true }
        profile: { type: [object, 'null'], additionalProperties: true }
        home: { type: string }
    ScreenName:
      type: string
      enum: [login, index, tutor, students, groups, group, invites, invite, lesson, homework, task, tutor-check, stats, bank, account, parent]
    ScreenState:
      type: object
      additionalProperties: true
...
  responses:
    Problem:
      description: Ошибка в формате RFC 9457 Problem Details
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    CommandSuccess:
      description: Команда выполнена
      content: { application/json: { schema: { type: object, additionalProperties: true } } }
    ScreenState:
      description: Специализированный bootstrap экрана без эталонных ответов
      headers: { ETag: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ScreenState' } } }
...